Attendance Management Revisions
The user or administrator should be able to add a record in the "Raw Time Logs" tab in case of a system down.

There are situations where employees still have to sign-in in a log book at times where the biometrics for the time in and out is down. So there should be a button, maybe call it "Add Time Log" or if you can think of a better button label, where if you click it then a modal would appear. Refer to the "raw_time_logs" table for the input fields that you need to include in the modal to add a time log. Maybe in the modal you should search first for the employee, select the employee, and then add a time log for that employee specifying the date and time of in and out. Feel free to do the best UX that you can think of.

No need to modify the not null constraint in the raw_time_logs table, it is only right that the admins or users can't delete or remove a registered time log. We need to have an "Add New Time Log" instead to comply with the professor's requirement.

Fixed the Generate CSV and PDF